backend-py/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.1

# Auth
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# HTTP + file handling
httpx==0.28.1
python-multipart==0.0.19
aiofiles==24.1.0

# Cloudinary
cloudinary==1.41.0

# Email
aiosmtplib==3.0.2

# Background workers
arq==0.26.1

# Observability
prometheus-client==0.21.1
structlog==24.4.0

# Rate limiting
slowapi==0.1.9

# OCR (routine scan)
pdfplumber==0.11.4
python-docx==1.1.2
easyocr==1.7.2
Pillow==11.1.0
numpy

# Spreadsheet parsing (marks upload)
openpyxl==3.1.5

# WebSocket (Socket.io compatible)
python-socketio==5.11.4
python-engineio==4.9.1

# Testing
pytest==8.3.5
pytest-asyncio==0.24.0
httpx==0.28.1
